pyproject.toml
[project]
name = "polito-task-arithmetic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "open-clip-torch>=2.29.0",
    "scipy>=1.14.1",
    "tqdm>=4.67.1",
]
